requirements.txt
python-dateutil==2.8.1
requests==2.28.1
beautifulsoup4==4.9.3